Split src/runs/spawn.ts into spawn/ subdirectory (warren-f71c / pl-9088 step 6)
Mirror the reap/ and stream/ layout (mx-5009ad family). The 882-line
src/runs/spawn.ts and 1344-line spawn.test.ts are decomposed into
per-concern modules under src/runs/spawn/, each with a matching test
file and a shared test-helpers.ts.

Modules:
  - dispatch.ts        — spawnRun orchestrator + composers + rollback
  - plot-append.ts     — run_dispatched Plot append (warren-e848)
  - seed-extensions.ts — post-dispatch sd update --extensions (pl-bb70)
  - agent-cache.ts     — readCachedAgent + override resolution
  - types.ts           — SpawnRunInput / SpawnRunResult / appender shapes
  - util.ts            — shared formatError
  - index.ts           — public re-exports

Tests (per source module, plus shared test-helpers.ts):
  - dispatch.test.ts        (validation + e2e + placement + rollback
                             + burrow_config/runtime/metadata
                             + composeDispatchPrompt)
  - refresh.test.ts         (project refresh + branch composition;
                             extracted to keep dispatch.test.ts ≤ 500 lines)
  - plot-append.test.ts     (plotId gating + PLOT env + run_dispatched)
  - seed-extensions.test.ts (seed extension write fire-and-log)
  - agent-cache.test.ts     (readCachedAgent + R-03 tier + provider/model)

Drops biome.json overrides for spawn.test.ts (noExcessiveLinesPerFunction)
and spawn.ts + spawn.test.ts (noExcessiveCognitiveComplexity); removes
both entries from scripts/file-size-budgets.json. test-helpers'
makeBurrowClient.fetchImpl exceeded cognitive complexity 15 inline, so
defaultBurrow/defaultBurrowRun/routeBurrowFetch were hoisted as
module-private helpers — the ratchet only goes down (AGENTS.md), no new
override was added.

src/runs/index.ts and src/runs/interactive.ts re-point to
./spawn/index.ts. Stale doc-comment references to src/runs/spawn.ts in
src/plan-runs/{errors,plot-appender}.ts, src/plots/{aggregate,id-validator}.ts,
src/seeds-cli/warren-extensions.ts, src/server/handlers/plan-runs.ts,
and src/runs/interactive.ts are updated to point at the new submodules.

bun run check:all → green (47 tests pass across the five new files).
