+++ .mulch/expertise/conventions.jsonl
@@ -3,3 +3,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T04:35:59.215Z","evidence":{"seeds":"warren-4553","commit":"fab6ba6fd2c479b5d490865aaf0061e0adea55b9"},"dir_anchors":["scripts"],"content":"File-size guard (warren-4553): per-file line cap enforced by scripts/check-file-sizes.ts (run via 'bun run check:file-sizes' in CI). Default threshold = 500 lines for .ts/.tsx under src/ and scripts/ (excludes src/ui, __golden__). Files exceeding 500 at the time of introduction are grandfathered in scripts/file-size-budgets.json at their commit-time line count — the ratchet only goes down, never raise a budget. Companion: biome.json's noExcessiveLinesPerFunction rule set to maxLines=500 with overrides disabling it for 8 baseline files (src/runs/reap.ts plus 7 test files). Refactoring those is tracked under warren-1b72.","id":"mx-5009ad"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T04:41:56.457Z","evidence":{"commit":"37675fc225a5ca4683db1dc91054f96c8e347515"},"relates_to":["mx-5009ad"],"content":"Debt-marker scanner (warren-7f2b / pl-7b06 step 8): scripts/check-debt-markers.ts (run via 'bun run check:debt-markers' and a CI step in .github/workflows/ci.yml, after check:file-sizes, before bun test) walks .ts/.tsx under src/ and scripts/ (excluding src/ui/, __golden__/, node_modules/) and fails on any TODO/FIXME/HACK/XXX marker whose line does not also contain a tracker reference: \\b(?:warren|pl|mx)-[0-9a-f]+\\b, #<digits>, or http(s) URL. Pattern \\b(TODO|FIXME|HACK|XXX)\\b deliberately does NOT match placeholder strings like 'pl-XXXX' (no word boundary inside XXXX). Grandfather list at scripts/debt-marker-allowlist.json holds 'path:line' strings; ratchet only goes down (stale entries that no longer match an untracked marker fail CI). The scanner and its test file (scripts/check-debt-markers.{ts,test.ts}) are in EXCLUDE_PATH_EXACT to avoid self-reference false positives. Companion to check-file-sizes.ts (mx-5009ad); same ergonomics and same scan roots.","id":"mx-02adb5"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T05:33:19.009Z","evidence":{"seeds":"warren-e5fb","commit":"d4ca38c864e4e99620604085ae9ed82f52452756"},"relates_to":["mx-fc83d2"],"content":"Automated doc-generation (warren-e5fb / pl-7b06 step 14): scripts/generate-docs.ts extracts the ROUTE_TABLE array from src/server/handlers.ts via brace-counting regex parsing and renders docs/http-api.md grouped by first path segment, preserving leading // ordering-caveat comments as the Notes column. Wired as 'bun run gen:docs' (write) and 'bun run gen:docs:check' (CI-fail on drift); check mode is part of 'bun run check:all'. Why not typedoc: HTTP routes are the meaningful contract; internal TS shapes churn too often. Regex parses the source string directly to avoid importing handlers.ts (heavy boot deps). After editing routes, regenerate the doc and commit.","id":"mx-22d123"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-27T21:00:23.033Z","evidence":{"commit":"22d16f5d473216009b7222cb07f1e7782418a98e"},"relates_to":["mx-5009ad"],"dir_anchors":["src/runs"],"content":"src/runs/spawn/ subdirectory layout (warren-f71c / pl-9088 step 6) follows the mx-5009ad split-by-concern shape established by reap/ and stream/. Modules: dispatch.ts (spawnRun orchestrator + provisionBurrow/dispatchRun/composeRunEnv/composeBurrowMetadata/composeDispatchPrompt/rollback), plot-append.ts (run_dispatched Plot appender + DEFAULT_DISPATCHER_HANDLE + resolveDispatcherHandle + defaultPlotAppender), seed-extensions.ts (pl-bb70 sd update --extensions fire-and-log), agent-cache.ts (readCachedAgent + resolveOverride + readProjectDefaults), types.ts (SpawnRunInput/SpawnRunResult/SpawnPlotAppender), util.ts (shared formatError), index.ts re-exports the public surface. Test files match 1:1 (dispatch/plot-append/seed-extensions/agent-cache/refresh.test.ts) sharing test-helpers.ts (makeBurrowClient/makePool/setupRepos/makeAgentJson). dispatch.test.ts hit the 500-line file-size budget so refresh+branch tests were extracted to refresh.test.ts. test-helpers' makeBurrowClient.fetchImpl exceeded cognitive complexity 15 inline; refactored by hoisting defaultBurrow/defaultBurrowRun/routeBurrowFetch helpers (the ratchet only goes down per AGENTS.md, no override added). biome.json overrides for spawn.test.ts (noExcessiveLinesPerFunction) and spawn.ts/spawn.test.ts (noExcessiveCognitiveComplexity) all removed; scripts/file-size-budgets.json entries for both files removed. Cross-module importers (src/runs/index.ts, src/runs/interactive.ts) re-point to ./spawn/index.ts.","id":"mx-fb160c"}
