+++ src/seeds-cli/warren-extensions.ts
@@ -21,7 +21,7 @@
  * `WarrenTriggerKind` locks down the trigger-string proliferation called
  * out as risk #6 in the plan. Today `src/server/handlers.ts` (Run Now)
  * writes `"manual-trigger"` into the warren `runs.trigger` column and
- * `src/runs/spawn.ts` defaults to `"manual"`; the enum here is the
+ * `src/runs/spawn/dispatch.ts` defaults to `"manual"`; the enum here is the
  * downstream-stable contract that step 4 will reconcile callers onto.
  *
  * The schema is `.strict()` for writes — unknown keys would silently
